adjust strategy for node
