+++ bower.json
@@ -1,6 +1,6 @@
 {
   "name": "fmd.js",
-  "version": "1.0.2",
+  "version": "1.1.0",
   "main": "dist/fmd.js",
   "ignore": [
     ".*",
