+++ lib/fmd.js
@@ -1,7 +1,7 @@
 /**
  * @module fmd.js in node
  * @author Edgar
- * @date 170213
+ * @date 170214
  * */
 
 
@@ -15,12 +15,9 @@ require('../src/injector/alias');
 require('../src/loader/resolve');
 require('../src/loader/assets');
 
-require('./loader/id2uri');
+require('./loader/id2url');
 require('./loader/noderequire');
 
 fmd.version = require('../package.json').version;
-fmd.config({
-    basePath: process.cwd()
-});
 
 module.exports = fmd;
