+++ lib/loader/noderequire.js
@@ -1,8 +1,8 @@
 /**
  * @module fmd/noderequire
  * @author Edgar <mail@edgar.im>
- * @version v0.1
- * @date 170104
+ * @version v0.2
+ * @date 170214
  * */
 
 
@@ -16,7 +16,7 @@ fmd( 'noderequire', ['config', 'module', 'assets'],
 
             var asset = assets.make( id, mod );
 
-            Module.has( asset.id ) || require( asset.uri );
+            Module.has( asset.id ) || require( asset.url );
 
             return Module.require( asset.id );
         };
