+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "fmd.js",
-  "version": "1.0.2",
+  "version": "1.1.0",
   "description": "another module writing system",
   "keywords": ["FMD","module","fmd.js","AMD","loader"],
   "homepage": "http://fmdjs.org/",
