--- test/specs/console/runner.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="utf-8"/>
-        <title>fmd/console - fmd.js Specs Runner - fmd.js单元测试</title>
-        <link rel="stylesheet" href="/test/resource/jasmine-1.3.1.css"/>
-        <script src="/test/resource/jasmine-1.3.1.js"></script>
-        <script src="/test/resource/jasmine-html-1.3.1.js"></script>
-
-        <!-- include source files here... -->
-        <script src="/src/boot.js"></script>
-        <script src="/src/utils/lang.js"></script>
-        <script src="/src/helper/console.js"></script>
-
-
-        <!-- include spec files here... -->
-        <script src="/test/specs/console/console.js"></script>
-
-        <script src="/test/resource/jasmine-runner-1.3.1.js"></script>
-    </head>
-
-    <body>
-    </body>
-</html>
